Clear the RunSXU flag between functions.
This was being set but not cleared before, so the effect was that whenever one
function needed its control flow preserved (because it contained calls,
generally) all of the functions after it would also be emitted with what was
left of their control flow intact whether they needed to be or not. This is
fixed by just clearing the flag after processing each function.
Old/split llvm commit ID: 9de938b438ee8f0cc827209fe59e8ea5cae2e97f