Remove extra dependency of recorder test
Recorder test had a dependency on the recorder (subject of testing),
but at a time we configure the target does not yet exists (or it is
disabled by cmake option). This produced a warning for CMake 2.8, but
for CMake 3.4 it is an error.

Removed this dependency to fix the build.

Change-Id: I0528667170f133fa3e4b05bea60f535f1208fe91