[SPMDization]
In SPMDization, the transformation of the condition of the loop generates always unsigned lt or unsigned gt.
The fix should be to generate signed or unsigned comparisons depending on the type of init and bound of the loop. However, LLVM does not distinguish between signed and unsigned integers. So, AFAIK, there is no way to check the sign of the integers.
This commit is a hack where I always generate signed comparisons (assuming that upper bound will never reach MAX_UINT)
This will work for now until I find a way to properly generate the appropriate condition
Old/split llvm commit ID: 3f5d31828cdeea2e9e352442e3cad9c001f767d8