cmplrs-46533 - add thread_private to omp threadprivate local static variables
Clang codegen uses a separate path for local statics so we need to set the
thread_private attribute there too.

Note we could instead do this during emission of the
OMPThreadPrivateDecl but these are ignored when not at the top level.  We will
likely have to revisit this to implement it for the C++ cases but this is
progress.

Change-Id: I1535015edc1e2698382d36f8bc87c53412fdbde9
Old/split clang commit ID: b9567de9b8579f3213b242aa7a4ce44b0700bc30