PipeOrdering: skip loops with #pragma unroll
When FPGA pipes or channels are used withing a loop, loop iteration
executes in a special order across work-items: loop iteration 0 of
each work-item in a work-group executes before iteration 1 of each
work-item in a work-group, and so on.

When #pragma unroll is used on a loop, it states that the loop can
be (and should be) unrolled without affecting a correctness of
execution. Unrolled loop may not conform to the FPGA loop
ordering (because it is no longer a loop), so for this we don't need a
barrier.

Omitting the barrier() for #pragma unroll is critical, because
otherwise it prevents optimizer from unrolling the loop (barrier() has
noduplicate attribute).

Change-Id: I605411d30a8b74b08f810645247070a4202dd08d