Tune the Optimizer flow to get rid of legacy meaning of HasGatherScatter flag.
Historically HasGatherScatter appeared after renaming of isKNC()
( http://cvcc-ubu-01.iil.intel.com/cr/r/3430/ )
Now we have to make clear distinction between AVX512
with and without Prefetch (SKX vs KNL), and not support KNC at all.

This patch makes the first attempt to unify the opt flow for Full Core
(SKX) and maintain some distinction of KNL (mainly for Prefetch).

There's still a couple of moments that are hard to make a decision
upon atm, like createReduceAlignmentPass and createPreLegalizeBoolsPass,
that are not being run on AVX-512 due to the historic meaning again.

Change-Id: I37da2e2219cd6a714c93cf1d0cb5909bac1bd053