Disable vectorization for kernels with FPGA features
This patch disables vectorization for "single work-item" kernels, which
can be vectorized by vector code never will be used due to the nature of
execution model.

We also disable vectorization for the kernels using channels/pipes as
read/write built-in functions for these data types are implemented with
caching. Unfortunately PipeSupport pass adds a loop directly to the user
call, which body *must* be executed atomically by each work-item (i.e.
we can't vectorize loop body).

TODO: consider resolution of pipe built-ins after vectorizer as it's
done for the rest of the built-in functions. This might allow us to
enable vectorization for kernels with pipe/channels.

Change-Id: Ied602083c6a7a36b8c664f6ffa2c3869e672c9fd