Add a side data table for tracking LIC information.
At present, the information in this data table isn't actually used anywhere yet.
The intent is to replace the INIT* instructions with the init vector, and to
use the register names in the very near future.

One caveat that's not well-solved yet is that LLVM may be prone to cloning
virtual registers and dropping them without maintaining information.
Old/split llvm commit ID: 7b248464e2164469367eb39f291ffbc5da677572