Refactor PipeSupport Pass to make possible processing of non-optimized IR
PipeSupport handles flush insertion which prevents deadlocks during
communication over pipes. It requires to know which pipes are used in the
current function. To figure it out PipeSupport tried to map builtin calls and
pipes used by these calls in compile time.

We can't do this mapping in general (for example, in non-optimized IR we
can have non-constant GEPs to arrays of pipes) in compile time.

This patch changes flush insertion behavior to rely on information we
can get in runtime. To support this PipeSupport inserts
'__store_pipe_use()' built-in calls before every pipe builtin call to store
used pipes in one of 2 arrays (the first one to store pipes with read access
and the second - with write access). So, '__flush_read_pipe_array()' and
'__flush_write_pipe_array()' will flush only pipes we already stored in arrays.

Change-Id: I17c345b1bf2e9398be9a030f362857ea8e9cb3fb