Make one of subtests of test_steps.py as expected to fail for GDB
'finish' command may work incorrectly in GDB/LLDB in some cases (e.g.
when it is done from a function which doesn't return any value or a function
called from a kernel).

It is expected that after 'finish' command a debugger stops on a line
just after a call of a function we've finished. But in some cases it
stops on the same line as the call.

The root case of the issue is unknown at the moment.
TODO: CORC-2448 Investigate and fix incorrect behaviour of 'finish' command.

Change-Id: I98b22d88d9ed648d21db4ad9759fb80c3c121063