Implementation of dynamic_cast calls optimization pass
When a destination class of a dynamic_cast call is a leaf class with
RTTI, i.e. there are no descedants with RTTI information of this class
in the whole program then we could make the following transformation
depending on the dynamic_cast hint:
  1. Negative hint.
    If we only need to know that dynamic_cast is succesfull or not then
    replace this test with comparison of pointers to type_info.
  2. Non-negative hint.
    Calculate the result of a dynamic_cast and replace all uses of
    the dynamic_cast.
Analysis of the hierarchy is based on type_info initializers. Linkage
type is used to be sure that there aren't any descedants with RTTI
information outside of LTO unit.

Transformation is done in case of Itanium ABI as it requires that:
  1. Every virtual table shall contain one entry that is a pointer to
    an object derived from std::type_info. This entry is located at the
    word preceding the location pointed to by the virtual pointer
    (i.e., entry "-1").
  2. The Itanium ABI specifies that that two type_info pointers point
    to equivalent type descriptions if and only if the pointers are
    equal, with one exception: if the type is an incomplete class type
    or a (possibly indirect) pointer to one. However, the language's
    built-in uses of RTTI generally require class types to be complete,
    even when manipulating pointers to those class types. That is why we
    could rely on address equality tests in case of dealing with
    dynamic_cast.

Change-Id: I530b13cf300b7c00457106cdc3df2da65a1e2d7e
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/107000
Reviewed-by: Gainullin, Artur <artur.gainullin@intel.com>
Tested-by: Gainullin, Artur <artur.gainullin@intel.com>
Old/split llvm commit ID: 46d2f37d854cc868ffd56e057b7381fd9c5b059f