[JIT Save/Load] Implement StripIntelIP pass.
Due to ICL policies it is not allowed to expose optimized
LLVM IR through any API or option in case Intel Proprietary
optimization have been performed. JIT Save/Load feature creates
an exposure point as JIT binary ELF has a section containing
optimized code.

At the same time it is not possible to exclude optimized
Module from the saved binary for 2 reasons:

* Internal OpenCL design that requires certain Metadata
  to be present in the Module to support OpenCL API queries.
  Such Metadata can only appear in middle-end.

  One example could be !opencl.global_variable_total_size
  supporting clGetProgramBuildInfo with
  CL_PROGRAM_BUILD_GLOBAL_VARIABLE_TOTAL_SIZE.

* MC JIT API that OpenCL RT uses requires to provide
  optimized Module to instantiate it.

Whilst the 2nd point can be resolved by turning to ORC JIT
that provides API not requiring a Module to be present,
the cost of such transition is estimated to be high in terms of
changing the existing code.

Although ORC JIT provides and adapter to the MC JIT API
for smooth transition, the APIs and containters for working
with precompiled objects have been redesigned dramatically.

Having said that the design decision is to exclude
as much detail from the optimized Module in a way that
does not intervene with mechanics of MC JIT, most
importantly the code content. For example,
exported functions, global ctor/dtor have to have
declarations in the Module, other internal functions
can be removed.

In fact, the online compilation through ORC JIT adapter
has been prototyped and the result is successful at least
from functional point of view.

Change-Id: Ie343b4e1539cd32ef6a731691da3c9a18299fb87