Fix mangling for arguments with private address space
In case of mangling/demangling of a pointer type we can treat as
substitutable entity these three type:
* a base pointee type, if it's substitutable (i.e, clk_event_t, float3, etc),
* a qualified pointee type, always when qualifiers exist (i.e. const float3),
* the whole pointer type (i.e. const float3 *).

Private address space is taken into account to define substitution
candidates now. So, in case of mangling/demangling of a pointer type we
always have a 'qualified' pointee type, which we should treat as substitutable
entity.

The patch:
* removes any checks before adding a qualified pointee type as a
substitution candidate, since we should always add it.
* make private address space is an explicit attribute for pointer types.

Change-Id: Iefb5f9f046d75c8ef40346949412a04ed1ec195d