Move DeadArgumentElimination pass
If a function has internal linkage type this pass can eliminate one or
even more arguments in a function call. Due to VPO passes are split in the
optimizer that will lead to a mismatch between number of parameters in
the function callee and its vectorized form. Therefore, this pass should be
launched before VPO.

Change-Id: I3eba53819867f80ee9f4b8f4f2a9024f965eb369