Don't give up on HIR Vec Dir insertion due to @llvm.assume calls.
It's not added to isTriviallyVectorizable check because it in fact is not - and
should be scalarized (yet to be done in VPlan Vectorizer). However, that should
not affect vectorizer's legality decision in any way.

Note, that community LV follows the same approach for handling the assumes - see
rL219741.

Change-Id: I0931bf8bf04ac847b6b1f60dd1e5bc68ccdefe2b
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/111980
Reviewed-by: Elovikov, Andrei <andrei.elovikov@intel.com>
Tested-by: Elovikov, Andrei <andrei.elovikov@intel.com>
Old/split llvm commit ID: 1e55ee0e0233eaafa9f8bb677bb3b3104314aacd