Adjust the register info to avoid attempting to allocate in *I0 classes.
This was reported as CMPLRS-48413 in Jira. Effectively, when choosing registers,
instruction selection attempts to find a constrained subclass that satisfies
all of the required constraints. In instruction selection, you get different
subclasses when one value is an I* constraint and the other is an RI*
constraint.

Since there is very little distinction between our different register classes,
the ties were broken by name. This caused the RI0 class to be chosen, and as
that is not allocatable, an assertion fired. What this change does is to
modify the spill sizes in such a way that breaking ties by name is not
necessary, and thus the RI1 class is chosen as the common subclass instead.
Old/split llvm commit ID: 392d863c98d1c49058095d5adf0b57e31a321d64