CSA ILPL: do not assume deep LICs
ILPL code generation has been improved so that minimum LIC depths are
specified for backedges. This ensures sufficient buffering in the loop
and avoids hangs even if we do not assume deep LICs.

E.g., ILPL should work even if the simulated program is run as:

    CSASIM_LIC_DEPTH_DEFAULT=3 CSASIM_LIC_DEPTH_DEFAULT_NARROW=3 \
    CSASIM_LIC_DEPTH_DEFAULT_MEZ=3 ./program ...

Another requirement to avoid hangs is to avoid the scenario where the pickany0
decides to admit another instance of the inner loop based on a particular
cohort while another cohort is blocked due to insufficient buffering /outside
of/ the inner loop pipeline, resulting in deadlock. This can be solved
either by additional buffering outside of the pipeline or by having the
pickany0 wait for the "all0" of new cohorts. This commit introduces the
latter and enables it by default.
Old/split llvm commit ID: fb608c3e6be8531e3fa6e279f0cd372c25d28fed