Move VPInstruction creation in plain CFG to decomposer.
This patch is a code movement and it's expected to be NFC. It moves the
responsibility of creating VPInstructions during the plain CFG construction
to the decomposer. Moving all the VPInstruction creation stuff to the
decomposer is reasonable because the decomposer will be creating VPInstructions
as a result of decomposition. This reduces the coupling of PlainCFGBuilderHIR
and VPDecomposerHIR. PlainCFBuilderHIR will continue with the responsibility of
creating VPBBs and will invoke the decomposer as an utility to create
(decomposed, in the future) VPInstructions.

Code in VPlanDecomposerHIR.cpp/.h is mainly from VPlanHCFGBuilderHIR.cpp.
They wouldn't need a throughout review.

Change-Id: Ibc1050cf0c6e6921d0212f7d975ef66477b7384e
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/125845
Reviewed-by: Caballero, Diego <diego.caballero@intel.com>
Tested-by: Caballero, Diego <diego.caballero@intel.com>
Old/split llvm commit ID: 4d734ea03c8b26a634308b6a02c35c1b0314e7f4