Add unit tests to handle alloca i1 in IR
LLVM IR can contain alloca instructions with i1 type. For x86 clang doesn't
generate such allocas - boolean values are being transformed to i8/i32 type. But
it appears to be that some OpenCL passes (read: ChannelPipeTransformation) can
add 'alloca i1' instruction. So the purpose of this patch is to verify
if we can handle this alloca appearing in IR.

The new tests checks validness of i1 type allocas in components where
the problem appeared at the first time.

Change-Id: Iac2d601f620987b5324baaae006ffb46eb826c1e
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/113499
Reviewed-by: Tyurin, Evgeniy <evgeniy.tyurin@intel.com>
Tested-by: Sidorov, Dmitry <dmitry.sidorov@intel.com>