[VPlan] Decompose HIR terminal RegDDRefs in VPlan.
This is the first big part of the decomposition algorithm which
decomposes RegDDRefs that are terminal. It's implemented in the
VPlanDecomposerHIR and it's integrated with the plain CFG construction.

TODOs:
- Decompose select instructions.
- Decompose HLIfs with multiple predicates.
- Decompose memory references.

KNOWN ISSUES: We are creating redundant VPInstructions for external
definitions. We need to introduce the proper VPlan representation for
them.

Change-Id: I2e765f409ad34246049cd9e0fbabc59a0ba33e2b
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/125847
Reviewed-by: Caballero, Diego <diego.caballero@intel.com>
Tested-by: Caballero, Diego <diego.caballero@intel.com>
Old/split llvm commit ID: 2bd2a23d0331a48354552bea5007c9c3451220c7