Add an utility function to detect images in the module
This patch moves functionality for detecting images from frontend
(clang) to backend. This is a part of work on removing
'opencl.used.optional.core.features' and 'opencl.used.extensions'
metadata generation from frontend.

I guess there are better approaches to do the things but we have only
one use of this information and it is for 'partially' (not officially)
supported device.

Change-Id: Ia730170e4bf61d6b734ce6d8ff27d346f697e3c1
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/114346
Reviewed-by: Bessonova, Kristina <kristina.bessonova@intel.com>
Tested-by: Bessonova, Kristina <kristina.bessonova@intel.com>