[SATest] Change the way of kernels lookup in SATest
SATest tries to find kernels in the module by 'opencl.kernels' metadata.
Since the frontend doesn't generate this metadata anymore,
so we can't assume that it is in the module and it has the correct
format, so lets look it up as a regular funtion in the module and
check that it has 'SPIR_KERNEL' calling convention.

Change-Id: I0f067477f22313df7c16db7e32abd15f96af2aae
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/114607
Reviewed-by: Tyurin, Evgeniy <evgeniy.tyurin@intel.com>
Tested-by: Bessonova, Kristina <kristina.bessonova@intel.com>