Add compiler tests to handle alloca i1 in IR
LLVM IR can contain alloca instructions with i1 type. For x86 clang doesn't
generate such allocas - boolean values are being transformed to i8/i32 type.
But it appears to be that some OpenCL passes (read: ChannelPipeTransformation)
can add 'alloca i1' instruction. So the purpose of this patch is to verify
if we can handle this alloca appearing in IR.

The new tests check validness of i1 type allocas in via running SATest.

Change-Id: Id188c8d2a27edd2639cce630cf2f76926a49305d
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/113976
Reviewed-by: Sachkov, Alexey <alexey.sachkov@intel.com>
Tested-by: Sidorov, Dmitry <dmitry.sidorov@intel.com>
Reviewed-by: Tyurin, Evgeniy <evgeniy.tyurin@intel.com>