Add optimization remarks for the streaming memory conversion pass.
With this change, it is possible to get notification of when streaming memory
conversion was enabled or when it failed to be enabled by using a command line
option:

$ clang -fopenmp -fopenmp-targets=csa redot.c -o redot -Rpass=.*
redot.c:9:20: remark: converted to streaming memory reference
      [-Rpass=csa-streamem]
      local_sum += vals[i];
                   ^

You need to add -Rpass=csa-streamem to see where the pass kicks in, and you
need -Rpass-missed=csa-streamem to see optimization failure remarks:

$ csa-clang -S -O3 foo.c -g -Rpass-missed=csa.*
foo.c:3:12: remark: streaming memory conversion failed: memory ordering tokens
      are not loop-invariant [-Rpass-missed=csa-streamem]
    a[i] = b[i];
           ^

You can get similar information if you use llc with -pass-remark=.* and
-pass-remark-missed=.* command line options. You must have the line information
embedded in the file (i.e., compile with -g) to have the result be useful.

Not all failure reasons are reported (indeed, the case where the initial
pattern match failed to match can't be reported), and there is a known case
where a spurious MOV causes the code to believe that the ordering tokens are
not loop-invariant when they in fact are. You can still get more precise
information with -debug-only=csa-streamem, but the reporting here is more
accessible to end-application developers.

Suggestions for clarifying the message are welcome.
Old/split llvm commit ID: 93d6df596cf08c224450a87f69051a943478d729