Add compiler tests to handle alloca of arbitrary integer type in IR
LLVM supports an arbitrary bit width for the integer type desired. Any
bit width from 1 bit to (2e23 - 1) can be specified. So the purpose of
this patch is to verify if the copiler can handle allocas of arbitrary
integers appearing in IR.

The new tests check validness of i239 type allocas in via running
SATest.

Change-Id: I1de3fab3d8f3ac1bae43d9584ef8a43e7ebce87a
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/114366
Reviewed-by: Sachkov, Alexey <alexey.sachkov@intel.com>
Tested-by: Sidorov, Dmitry <dmitry.sidorov@intel.com>
Reviewed-by: Tyurin, Evgeniy <evgeniy.tyurin@intel.com>
Reviewed-by: Savonichev, Andrew <andrew.savonichev@intel.com>