Fix streaming memory miscompilation when do/while loops are involved.
The solution here is to map the length parameter with a trip count that is at
least 1, even if the condition would fail. The resulting code is unnecessary if
the loop is already guarded in a zero-trip-count check; the plan is to convert
such loops to a SEQ* operator instead of a SEQOT* operator, so special-casing
such situations is deferred for the moment.

Note that the generated length conditions are checked for uniqueness of code.
We should unify such CSE-esque code for other transformations as well.
Old/split llvm commit ID: 813e12e3627d79772efee9c067a2d88ac9c6ba74