[SATest] Make reference built-ins as shared library
Currenly reference built-ins version is built as static library and
linked to SATest executable. In the case the built-ins aren't visible
for dlsym which used by LLVM to find BIs implementation for the interpreter.

To fix it this patch makes shared library from reference built-ins and
links it with SATest executable.

To avoid any ambiguities with names, the library with reference
built-ins is renamed to OCLBuiltinsRef.

Note: This approach hasn't been tested on Windows, since SATest can't be
built on Windows (SATest's build is turned off by defalut, no ICC on
Windows to build IMathLibd, etc).

Change-Id: I04909fc9474a567371835fe7df675a8e0e74e2e4
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/117235
Reviewed-by: Tyurin, Evgeniy <evgeniy.tyurin@intel.com>
Tested-by: Bessonova, Kristina <kristina.bessonova@intel.com>