CSA ILPL: handle the case where the inner loop has no outputs at all
In this case we don't need to reorder anything, but we still need a
mechanism to control the number of cohorts allowed in the loop. For now,
because it's easy, we do this with a completion1's indices. An
alternative would be something like the previous pre-initialized
"semaphore" LICs.
Old/split llvm commit ID: 9e41b55d7bf4215cea5bf57c262805a7985a3d8e