[FPGA] Implemented __builtin_fpga_reg built-in
Purpose: PSG has decided to add a new compiler built-in function for
our 18.0 release, to enable low level optimization on our newest device
family.  The SSG FPGA emulator will need to accept the function in 18.0
and onward.  It should be implemented as an assignment within the
emulator.

Motivation: The function is __builtin_fpga_reg( operand  ), and is legal only
in the kernel language.  It takes a single argument/operand which can
be a first class lvalue (not a channel for example) or an rvalue.  In
FPGA hardware, it semantically means that at least one hardware
register will be inserted between the operand the output of the
built-in.  The effect of this is observable only through the frequency
that a design operates at in FPGA hardware, so it lowers to an
assignment in all other contexts.  It is safe to optimize away during
CPU compilation. Only very advanced users should make use of it.

Limitations: This built-in works in HLS only for POD types.

Implementation strategy: We generated two intrinsics:
* T __builtin_fpga_reg(T) - it used for all types except structures/unions.
* void __builtin_fpga_reg(T* ret, T*arg) - it used for structures/unions.

In Sema for __builtin_fpga_reg built-in we change return type to type of the
argument. After that in CodeGen we generate one of two intrinsics
depends on type of the arugment.

Depends-On: I071a903024e531b4aa226b91227f6debfc6020bf
Change-Id: I60c69f9054093255197c4dbe13e39fa06b2f1e20
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/118344
Reviewed-by: Churaev, Egor <egor.churaev@intel.com>
Tested-by: Churaev, Egor <egor.churaev@intel.com>
Old/split clang commit ID: ae08edf7f1a425be27aed2572c0db3ab5c51d9a9