[SPIRV] Generate OpControlBarrier for Subgroup scope for OpenCL 1.2/2.0
'cl_intel_subgroup' extension allows to use sub_group_barrier() built-in
with OpenCL 1.2/2.0. It requires to translate this built-in to
reqular (core) SPIR-V instruction OpControlBarrier.

Note: currently clag generates nothing representing vendor extensions
like 'cl_intel_subgroup'. So, SPIR-V translator doesn't check that the
extensions was enabled. It assumes that if a built-in or a type which
belong to one of supported extension used in the IR, this extensions is
enabled.

Change-Id: I707a48f5ea647289902bf26544c49044a48af5a3
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/119713
Reviewed-by: Bessonova, Kristina <kristina.bessonova@intel.com>
Tested-by: Bessonova, Kristina <kristina.bessonova@intel.com>
Old/split llvm commit ID: 628ae49d5a9bdd023ae20cc32b1751f1fbd8b014