Fix for clSetKernelArg.
When there is no local_mem_size attribute then value of localMemSize
should be 0. It means that size of local argument is not declared and
can be any.
There was no check on zero in clSetKernelArg function. So, the function
has returned CL_INVALID_ARG_SIZE for any local arguments except args
with local_mem_size attribute.

Change-Id: I398c651f5867bdea29320008e458da92d9d27169
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/125677
Reviewed-by: Tyurin, Evgeniy <evgeniy.tyurin@intel.com>
Reviewed-by: Romanov, Vlad <vlad.romanov@intel.com>
Tested-by: Churaev, Egor <egor.churaev@intel.com>