The changes here fix 'Old index not found' assert in CanonExpr::replaceBlob.
The issue was that when collecting the blob indices in a CanonExpr, we were
not making the collected indices unique. When a CanonExpr uses the same blob
multiple times, we were hitting the assert during vectorization when we try
to replace a blob with the vector equivalent after the first time.

The changes in HIR-CG broadcast an IV value when needed.

Change-Id: Id9121662118a6a456d8510f165f324fdd878b212
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/126148
Tested-by: zperf_checkin_xmain <lab_iclz@sceler38.sc.intel.com>
Reviewed-by: Chawla, Pankaj <pankaj.chawla@intel.com>
Tested-by: Guggilla, Satish <satish.guggilla@intel.com>
Reviewed-by: Kreitzer, David L <david.l.kreitzer@intel.com>
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/126589
Reviewed-by: Guggilla, Satish <satish.guggilla@intel.com>
Old/split llvm commit ID: ef53db55cce42331375c1e1e9518fa7860a92b1f