Build LIC groups and attach block frequency information.
For the moment, the output is disabled unless you explicitly request it. There
may be better ways to expose it, so I don't want to bake this in at the moment.

There remains an assertion if the block frequency is mixed up in the
calculations. The code should be robust enough at this point, that failure is
more likely to indicate bugs in dataflow conversion than the block frequency
or LIC grouping code, so the assertion remains.

Change-Id: I43ceadc53b3d686532fe42d52147bc07a53ea4e9
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/125503
Reviewed-by: Zakharin, Vyacheslav P <vyacheslav.p.zakharin@intel.com>
Tested-by: Cranmer, Joshua <joshua.cranmer@intel.com>
Old/split llvm commit ID: b5de57ca5949ce10c89f77c9fb3c0f38701d3ad3