Fix structure layouts with ap-ints.
AP-Ints were being converted to arrays in structures when they were
not the last item in the array.  This is because they were running afoul
of code intended to properly set the alignment for bitfields.

The fix to this is simply to ensure that the llvm alloca size of these
types matches the structure layout size.

Change-Id: I27403e10dfa6dc7d1518a740735fe04e6fa02ca6
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/128271
Reviewed-by: Andrews, Elizabeth <elizabeth.andrews@intel.com>
Reviewed-by: Felipe de Azevedo Piovezan <felipe.de.azevedo.piovezan@intel.com>
Reviewed-by: Rice, Michael P <michael.p.rice@intel.com>
Reviewed-by: Kreitzer, David L <david.l.kreitzer@intel.com>
Tested-by: Kreitzer, David L <david.l.kreitzer@intel.com>
Old/split clang commit ID: 20c7a48f0f3d506def4637d0830cbcc5b39fbe64