cmplrs-45956: implementation of #pragma nofusion
Pragma is a loop-hint pragma that goes on loop statements.
icc does not specify what kind of loop statement it needs to be on.
In xmain, we simply generate loop metadata for the loop it is on.

Change-Id: I9bdebc568bb877e250fd67e03cd93d37a8bebb7d
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/127682
Reviewed-by: Rao, Premanand M <premanand.m.rao@intel.com>
Tested-by: Rao, Premanand M <premanand.m.rao@intel.com>
Old/split clang commit ID: 013430590ddf994566a85d9c983a4b9769d1458b