Fix crash when assigning to an uapint 1
Since 'true' and 'false' constants are 1 bit values and bools are stored
in 8 bits, there is a hack in clang to make sure that this constant
value is zero extended to fill the boolean.  However, it detects this
simply by checking the size of the constant!  In our case, we have
legitimate 1 bit values, so only zero-extend if the size is at least 1.

Change-Id: Iab8ce255a84530ffc066b76587c1a79ecaf902c2
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/128272
Reviewed-by: Andrews, Elizabeth <elizabeth.andrews@intel.com>
Reviewed-by: Felipe de Azevedo Piovezan <felipe.de.azevedo.piovezan@intel.com>
Tested-by: Felipe de Azevedo Piovezan <felipe.de.azevedo.piovezan@intel.com>
Reviewed-by: Rice, Michael P <michael.p.rice@intel.com>
Reviewed-by: Kreitzer, David L <david.l.kreitzer@intel.com>
Old/split clang commit ID: b061830f2e85e2bde2906bd6bd47c50116d10b59