Actually delete fields in the dtrans DeleteFields pass
This change updates the DeleteField pass so that the populateTypes
function actually deletes the fields we have determined are not used.
GetElementPtr instructions that access the deleted fields (along with
their uses) are deleted during processFunction, and GEPs accessing
fields whose indices are changed by the field deletion are updated
during postprocessFunction.

This does not update the instructions that reference the size of the
structure, so the transformation does not produce correct code yet.

Change-Id: I05b97eb5b873c48b8d34599b889d37a99b494d0d
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/129487
Tested-by: xmain_checkin_sanity <lab_icls@sceler38.sc.intel.com>
Reviewed-by: Kaylor, Andrew <andrew.kaylor@intel.com>
Reviewed-by: Kreitzer, David L <david.l.kreitzer@intel.com>
Old/split llvm commit ID: 86d0929719686968a789fc049e54b764f462e970