[dtrans] Infer allocated types from what is stored there
Previously the local pointer analysis was able to infer the type of an
allocation from store instructions if the allocated object was the
value being stored. This change adds new logic to also infer the
allocated type from what is being stored to the allocated pointer.

This is needed for:

 CMPLRS-45540

Change-Id: Ifd4781b8b6abf2a88dd6d9b31eedb505c0a552b5
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/128769
Reviewed-by: Kaylor, Andrew <andrew.kaylor@intel.com>
Tested-by: Kaylor, Andrew <andrew.kaylor@intel.com>
Old/split llvm commit ID: 2000dafc670da09131f6e382cffda24ea19b37dc