CMPLRS-50588: Fix conversions from AP-Int to Bool
Conversions from an integer to a boolean have a special ICK type that
was not considered in the initial implementation. Thus, it assumed that
it was an Enum being converted to a bool.  This patch ensures that the
integer to bool conversion results in ICK_Boolean_Conversion, which
generates the proper IR.

Change-Id: I6e91cf7159a4db194d27c392333c37510b4d1642
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/130019
Reviewed-by: Andrews, Elizabeth <elizabeth.andrews@intel.com>
Reviewed-by: Rice, Michael P <michael.p.rice@intel.com>
Reviewed-by: Felipe de Azevedo Piovezan <felipe.de.azevedo.piovezan@intel.com>
Tested-by: Felipe de Azevedo Piovezan <felipe.de.azevedo.piovezan@intel.com>
Tested-by: xmain_checkin <lab_iclt@sceler38.sc.intel.com>
Reviewed-by: Kreitzer, David L <david.l.kreitzer@intel.com>
Old/split clang commit ID: c878848243a118d2154deca4033daa03379bc6ac