[dtrans] AOS-to-SOA support for creating peeled structure types and global variable of type.
This change implements the first part of the AOS-to-SOA transformation class that derives
from the DTransOptBase class. This version supports the methods needed to create and
populate the new peeled types. It does not support the IR instruction transformation
beyond that provided by the base class, and therefore can only be used on IR that
does not directly access the fields of the structure being converted in this version.

Change-Id: I12f54e78b7ef586a00c5554fedd6a2d486dc4926
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/130101
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Tested-by: xmain_checkin <lab_iclt@sceler38.sc.intel.com>
Reviewed-by: Kreitzer, David L <david.l.kreitzer@intel.com>
Old/split llvm commit ID: 97c4fad70b18c7da6e7caa9348abcf3e99e05de9