[dtrans] Detect non-div uses after ptr sub
Previously our analysis of pointer subtraction did not consider the uses
of the subtraction result. In order for pointer subtraction to be safe,
there must not be any uses other than division by the size of the
structure that was pointed to. We "look through" PHI nodes and selects
to find other uses.

If the pointers being subtracted were pointers-to-pointers, we do not
need to perform this check.

CMPLRS-45540

Change-Id: I5185926995c4a4f739595b9ef36443becc0bc568
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/130108
Reviewed-by: Kaylor, Andrew <andrew.kaylor@intel.com>
Tested-by: Kaylor, Andrew <andrew.kaylor@intel.com>
Old/split llvm commit ID: bacf853c26c4922cad7cd09d6fe8c2c0d359cdf6