No more Twine pre-computation in CSAReassocReduc.
This removes a pre-computed lic name prefix that this pass used to have
in favor of a helper lambda which calculates the prefixed lic names
internally. With the old approach some of the components of the twine
were apparently not surviving until the twine's use; rebuilding the
twine inside of a lambda each time fixes that.

Change-Id: I049a8089355defb68f8f5b6d3a5bd8595975639d
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/130217
Reviewed-by: Zakharin, Vyacheslav P <vyacheslav.p.zakharin@intel.com>
Tested-by: Woodworth, Daniel <daniel.woodworth@intel.com>
Old/split llvm commit ID: 094e65f50f44e04eb4e15762feb09dc88ee69547