Update size arguments after dtrans field deletion
This change continues the implementation of the DTrans DeleteField pass
adding the ability to update size arguments to allocation and memfunc
calls after fields have been deleted. The def-use chain of the size operands
are followed to handle the case where the actual argument to the call is
a value calculated from a constant in previous instructions.

This does not yet update the size operand in the case where pointer
subtraction is followed by a div by the size of a structure.

CMPLRS-45540

Change-Id: I29ca7bbb035d47f250b805b6428189ea8bde311a
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/130768
Reviewed-by: Kaylor, Andrew <andrew.kaylor@intel.com>
Tested-by: Kaylor, Andrew <andrew.kaylor@intel.com>
Old/split llvm commit ID: 4ff0bfaf03ea5452d4072fca219154a4daef3e0e