[dtrans] cmplrs-50640: Memcpy analysis needs to consider case with pointer to member of struct and array.
When checking for handled cases of memcpy/memmove operands, the analysis was dereferencing a nullptr
if the source operand was a member of an array aggregate being copied to an array aggregate of a
structure. This change marks the types as "unhandled use"for consistency with the way it would be
handled if the target structure were an array aggregate.

Change-Id: Iabca8fc2f01cd689a038f983bb4b14cb1774c90b
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/130534
Tested-by: xmain_checkin_sanity <lab_icls@sceler38.sc.intel.com>
Reviewed-by: Chupina, Olga <olga.chupina@intel.com>
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Reviewed-by: Kreitzer, David L <david.l.kreitzer@intel.com>
Old/split llvm commit ID: 18111015135ff18ac51fb45e0222f910c96ee8f6