Typedef 'long long' type in x86 intrinsics headers
Intrinsics headers are used for compilation of both C and OpenCL C
sources. In C 'long long' type is 64bit, while in OpenCL C it is
128bit.

Since intrinsics headers originally were implemented only for C, they
are not compatible with OpenCL C. This patch adds a typedef __i64 and
__u64 instead of 'long long'.

Original problem: some intrinsics which have 'long long' type were
incorrectly declared as 128bit, but definition was 64bit. Thus we had
a function cast in IR:

  %call.i.i = call <8 x i64> bitcast (<8 x i64> (i8, i64)* @_mm512_maskz_set1_epi64
      to <8 x i64> (i8, i128)*)(i8 zeroext %14, i128 4607182418800017408)

and CodeGen handled this badly.

This patch fixes Conformance1.2/Commonfns12 on Windows AVX512.

Change-Id: I0277bca7cd581b046e9f46d26761b6eb4e4ac133
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/130916
Reviewed-by: Tyurin, Evgeniy <evgeniy.tyurin@intel.com>
Tested-by: Savonichev, Andrew <andrew.savonichev@intel.com>