[dtrans] DTransOptBase class functionality to update CallInfo data structure during cloning/remapping
This change adds the updating of the CallInfo objects to occur prior to calling the
postProcessFunction() of the derived class so that when the post processing is looking
at a call, it can lookup the callinfo. Note: For cloned functions, the lookup needs
to be done on the cloned function, not the original instruction that will be deleted.

Change-Id: Icd4e5f0507968a1d3c380b4e32964b45f658cbab
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/131104
Tested-by: xmain_checkin <lab_iclt@sceler38.sc.intel.com>
Reviewed-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Reviewed-by: Kreitzer, David L <david.l.kreitzer@intel.com>
Old/split llvm commit ID: 09873133738fe8194340e0492ae4cfa04d798e8a