[VPO] Change the value of "QUAL.OMP.ORDERED" for ordered clause to 0.
This is for the 'ordered' clause without a parameter, such as:
  "#pragma omp [parallel] for ordered

For this, the value sent to the backend will be 0, whereas it used
to be 1. This is to differentiate between the parameterized
'ordered(n)' clause which represents a do-across loop nest. Here
'n' can never be 0 (according to OpenMP 5.0 spec).

The VPO change to handle this is here:
https://git-amr-2.devtools.intel.com/gerrit/#/c/131102/

Change-Id: Ic43ce1c67886b534de756efec7b580a72d116d1f
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/131301
Tested-by: Gaba, Abhinav <abhinav.gaba@intel.com>
Reviewed-by: Su, Ernesto <ernesto.su@intel.com>
Reviewed-by: Rice, Michael P <michael.p.rice@intel.com>
Reviewed-by: Kreitzer, David L <david.l.kreitzer@intel.com>
Old/split clang commit ID: 24bd95b3808b2ed8177f433f71ffd8aad64ee476