[dtrans] Add dtrans-analysis test for function bitcasts
Bitcasts of function pointers have an effect that is equivalent to
bitcasting each of the arguments to the call. This test case shows what
the safety conditions should be in the presence of a function bitcast.
However, the dtrans analysis does not currently support this so the test
is marked as XFAIL.

A future changeset will implement support for analyzing function
bitcasts and the XFAIL will be removed at that time.

CMPLRS-50759

Change-Id: I9fce2d8338a217acc182bcadcd8b2863d9d14b88
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/131743
Reviewed-by: Kaylor, Andrew <andrew.kaylor@intel.com>
Tested-by: Kaylor, Andrew <andrew.kaylor@intel.com>
Old/split llvm commit ID: d7f67e7684ef529bd44fb7d19aca40f55262428d