CORC-2969: Enable OpenCL 2.1 specification for all supported CPUs
There is a decision, to build OpenCL-RT with 'BUILD_EXPERIMENTAL_21'
cmake variable, which turn on "opencl-2_1" mode and make this mode
as default one. The fix prepares the source code according to
the decision. After fix promotion, QA-build toolls should be fixed as well.

Brief overview -
1. I replaced macro '-DBUILD_EXPERIMENTAL_21' by '-DBUILD_OPENCL_21'.
2. I replaced platform name "Experimental OpenCL 2.1 CPU Only Platform"
   by "Intel(R) OpenCL 2.1 CPU Only Platform"

All changes are because "opencl-2_1"  mode will be the default mode,
not an experimental one.

Change-Id: I56cc20702e54df663150972e7fdb4a54cd0cbbee
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/131860
Reviewed-by: Romanov, Vlad <vlad.romanov@intel.com>
Tested-by: Nazarov, Mikhail V <mikhail.v.nazarov@intel.com>