[dtrans] Update ptr sub size divisor following delete field
This change implements the code to update the constant by which a
pointer subtract result is divided after the delete field pass has
deleted a field from a structure for which pointers to the structure are
subtracted. The DTransAnalysis will have checked to make sure that all
uses of the pointer subtraction are divides by some multiple of the
structure size.

This change updates the DTransAnalysis so that it will no longer
accept pointer subtract results being passed through PHI nodes and
select instructions before being divided (since we can't easily know
that the other input values should experience the same division). The
analysis also now records the primary type of interest involved in
pointer subtraction instructions it has analyzed.

At this point the pointer subtraction type mapping is not updated when
types are replaced, so this version of the code relies on the original
mapping. This will not work for functions which must be cloned. There is
a comment in the new code noting the need to update this.

CMPLRS-45540

Change-Id: I12d3455b0e1ba335429068e7c231c29cb868f350
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/131097
Reviewed-by: Kaylor, Andrew <andrew.kaylor@intel.com>
Tested-by: Kaylor, Andrew <andrew.kaylor@intel.com>
Old/split llvm commit ID: 747f914db6bd321416c96b93a3bb3329665c4ee4